archive/requirements_new.txt
# Core Flask
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Limiter==3.10.0

# Security & Authentication
Werkzeug==3.0.3
Authlib==1.6.6
cryptography==46.0.3
python-dotenv==1.2.1

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Machine Learning
xgboost==3.1.2
scikit-learn==1.8.0
numpy==2.2.3
pandas==2.2.3
joblib==1.5.2
scipy==1.16.3
shap==0.50.0

# AI & Document Processing
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.187.0
google-auth==2.38.0
google-cloud-aiplatform==1.82.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.8
pdfminer.six==20251107
reportlab==4.4.6
pypdfium2==5.2.0

# Image Processing
Pillow==11.1.0
opencv-python==4.12.0.88
pytesseract==0.3.13

# Visualization
matplotlib==3.10.8
qrcode==8.2

# Task Scheduling
APScheduler==3.10.4

# Development Tools
Faker==38.2.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0

# Utilities
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2025.1
